pkg: pine64: linux-pine64: enable dm-thin-pool kernel module
This PR enables the dm-thin-pool kernel module which is
needed to run docker in the standard configuration.

To test: just install docker and start the docker.service.
If it fails, most likely the kernel module is missing. After
recompiling with this config enabled, it works.

I think this is something which should be enabled by default.
As far as I know manjaro supports docker out of the box on
the pinephone.

Signed-off-by: Danct12 <danct12@disroot.org>