WIP: rewrite build.sh without guzuta
WARNING: this is incompatible with guzuta, it may overwrite database
files generated by guzuta, do not attempt to use this with repositories
managed by guzuta. Hopefully this will be addressed by a future version
of the script.

This has the benefit of also being able to use makearmpkg from the
armutils-git package to crossbuild packages from an x86 machine.